+++ CHANGES.md
@@ -1,6 +1,8 @@
 # Changes
 
-## Version 1.0.0 (Not yet Released)
+## Version 1.0.1 (Not yet Released)
+
+## Version 1.0.0
 
 ## Version 1.0.0-beta7